+++ src/Facades/HelperCache.php
@@ -7,6 +7,9 @@ use Illuminate\Support\Facades\Facade;
 /**
  * @see \CodebarAg\LaravelDefault\Helpers\HelperCache
  *
+ * @method static int cacheForeverWithTags(array $tags, string $key, $store)
+ * @method static int cacheRememberWithTags(array $tags, string $key, int $time, $store)
+ * @method static int flushTags(array $tags)
  * @method static int inSeconds()
  */
 class HelperCache extends Facade